+++ Changelog.md
@@ -18,6 +18,7 @@
 - Fix: Breaking an Inventory Scanner/Laser or one of its field does not break all fields, if the respective config setting is lower than the amount of fields present
 - Fix: Breaking an Inventory Scanner Fields with the inventoryScannerRange config setting set to >=3 does not break and reinstate all fields properly
 - Fix: Players that are on the allowlist of a Security Camera, cannot toggle the camera's redstone output
+- Fix: Disabled option in laser blocks does not work properly
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------